+++ weave/legacy/weavify.py
@@ -2,8 +2,8 @@ import typing
 
 from weave.legacy import graph, op_args, val_const
 
-from . import errors, weave_internal
-from . import weave_types as types
+from .. import errors, weave_internal
+from .. import weave_types as types
 
 if typing.TYPE_CHECKING:
     from weave.legacy import op_def
