+++ weave/tests/legacy/test_weavify.py
@@ -2,8 +2,9 @@ import pytest
 
 from weave.legacy import graph, ops
 
-from ... import weave_internal, weavify
+from ... import weave_internal
 from ... import weave_types as types
+from ...legacy import weavify
 from .. import geom
